refactor: split workspace state
