+++ whitenoise-mac/Core/GroupImageSearch/GroupImageSearchError.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+enum GroupImageSearchError: LocalizedError {
+    case invalidURL
+    case invalidResponse
+    case requestFailed(statusCode: Int)
+
+    var errorDescription: String? {
+        switch self {
+        case .invalidURL:
+            return L10n.string("Could not build the image search URL.")
+        case .invalidResponse:
+            return L10n.string("The image search service returned an invalid response.")
+        case .requestFailed(let statusCode):
+            return String(format: L10n.string("Image search failed with HTTP status %d."), statusCode)
+        }
+    }
+}
