+++ whitenoise-mac/Core/GroupImageSearch/GroupImageSourceLoading.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+/// Fetches the bytes for an image the user picked out of search results.
+protocol GroupImageSourceLoading {
+    func data(for url: URL) async -> Data?
+}
