+++ whitenoise-mac/Core/GroupImageSearch/OpenverseGroupImageSearchClient.swift
@@ -1,0 +1,32 @@
+import Foundation
+
+struct OpenverseGroupImageSearchClient: GroupImageSearchClient, Sendable {
+    private let endpoint = URL(string: "https://api.openverse.org/v1/images/")!
+
+    func searchImages(query: String) async throws -> [GroupImageSearchResult] {
+        var components = URLComponents(url: endpoint, resolvingAgainstBaseURL: false)
+        components?.queryItems = [
+            URLQueryItem(name: "q", value: query),
+            // Openverse caps anonymous (unauthenticated) requests at page_size 20 and
+            // rejects anything larger with HTTP 401, so stay at the anonymous ceiling.
+            URLQueryItem(name: "page_size", value: "20"),
+            URLQueryItem(name: "mature", value: "false"),
+        ]
+
+        guard let url = components?.url else { throw GroupImageSearchError.invalidURL }
+
+        var request = URLRequest(url: url)
+        request.setValue("WhiteNoiseMac/1.0", forHTTPHeaderField: "User-Agent")
+
+        let (data, response) = try await URLSession.shared.data(for: request)
+        guard let httpResponse = response as? HTTPURLResponse else {
+            throw GroupImageSearchError.invalidResponse
+        }
+        guard (200..<300).contains(httpResponse.statusCode) else {
+            throw GroupImageSearchError.requestFailed(statusCode: httpResponse.statusCode)
+        }
+
+        let decoded = try JSONDecoder().decode(OpenverseImageSearchResponse.self, from: data)
+        return decoded.results.compactMap(\.groupImageSearchResult)
+    }
+}
