+++ whitenoise-mac/Core/GroupImageSearch/OpenverseImageSearchResponse.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct OpenverseImageSearchResponse: Decodable {
+    let results: [OpenverseImageRecord]
+}
