+++ whitenoise-mac/Core/GroupImageSearch/SecureGroupImageSourceLoader.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct SecureGroupImageSourceLoader: GroupImageSourceLoading {
+    func data(for url: URL) async -> Data? {
+        await RemoteImageLoader.shared.data(for: url)
+    }
+}
