+++ whitenoise-mac/Core/MarmotConcurrency.swift
@@ -2,7 +2,7 @@ import MarmotKit
 
 // UniFFI generates plain value records for these FFI types and declares no Sendable
 // conformances. WorkspaceState marshals them across its off-main FFI boundary (see
-// `runOffMain`); each is an immutable snapshot produced by a single Rust call, so an
+// `FFIExecutor.run`); each is an immutable snapshot produced by a single Rust call, so an
 // unchecked Sendable conformance is sound. `@retroactive` documents that the app — not
 // the generated module — vends the conformance, and silences the cross-module warning.
 extension TimelineMessageQueryFfi: @retroactive @unchecked Sendable {}
