+++ whitenoise-mac/Core/Notifications/MacLocalNotificationCenter.swift
@@ -1,0 +1,95 @@
+import Foundation
+import UserNotifications
+
+@MainActor
+final class MacLocalNotificationCenter: NSObject, LocalNotificationCenter, UNUserNotificationCenterDelegate {
+    private let center: UNUserNotificationCenter
+    private var responseHandler: (@MainActor ([String: String]) -> Void)?
+
+    init(center: UNUserNotificationCenter = .current()) {
+        self.center = center
+        super.init()
+        center.delegate = self
+    }
+
+    func authorizationStatus() async -> LocalNotificationAuthorizationStatus {
+        await currentSettings().authorizationStatus.localNotificationStatus
+    }
+
+    func requestAuthorization() async throws -> LocalNotificationAuthorizationStatus {
+        _ = try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Bool, Error>) in
+            center.requestAuthorization(options: [.alert, .sound, .badge]) { granted, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: granted)
+                }
+            }
+        }
+        return await authorizationStatus()
+    }
+
+    func post(_ notification: LocalNotificationRequest) async throws {
+        let content = UNMutableNotificationContent()
+        content.title = notification.title
+        content.body = notification.body
+        content.sound = .default
+        content.threadIdentifier = notification.threadIdentifier
+        content.userInfo = notification.userInfo
+
+        let request = UNNotificationRequest(
+            identifier: notification.identifier,
+            content: content,
+            trigger: nil
+        )
+
+        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
+            center.add(request) { error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume()
+                }
+            }
+        }
+    }
+
+    func setResponseHandler(_ handler: @escaping @MainActor ([String: String]) -> Void) {
+        responseHandler = handler
+    }
+
+    nonisolated func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        didReceive response: UNNotificationResponse,
+        withCompletionHandler completionHandler: @escaping () -> Void
+    ) {
+        let userInfo = response.notification.request.content.userInfo.reduce(into: [String: String]()) {
+            result, element in
+            guard let key = element.key as? String else { return }
+            if let value = element.value as? String {
+                result[key] = value
+            }
+        }
+
+        Task { @MainActor [weak self] in
+            self?.responseHandler?(userInfo)
+            completionHandler()
+        }
+    }
+
+    nonisolated func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        willPresent notification: UNNotification,
+        withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void
+    ) {
+        completionHandler([.banner, .list, .sound])
+    }
+
+    private func currentSettings() async -> UNNotificationSettings {
+        await withCheckedContinuation { continuation in
+            center.getNotificationSettings { settings in
+                continuation.resume(returning: settings)
+            }
+        }
+    }
+}
