+++ whitenoise-mac/Core/Notifications/UNAuthorizationStatus+LocalNotificationStatus.swift
@@ -1,0 +1,20 @@
+import UserNotifications
+
+extension UNAuthorizationStatus {
+    var localNotificationStatus: LocalNotificationAuthorizationStatus {
+        switch self {
+        case .notDetermined:
+            .notDetermined
+        case .denied:
+            .denied
+        case .authorized:
+            .authorized
+        case .provisional:
+            .provisional
+        case .ephemeral:
+            .ephemeral
+        @unknown default:
+            .denied
+        }
+    }
+}
