+++ whitenoise-mac/Core/WorkspaceState.swift
@@ -5,964 +5,6 @@ import Foundation
 import MarmotKit
 import Observation
 import SwiftUI
-import UserNotifications
-
-private func uniquedLastWins<Value, Key: Hashable>(
-    _ values: [Value],
-    by key: (Value) -> Key
-) -> [Value] {
-    var seenKeys = Set<Key>()
-    var result: [Value] = []
-    result.reserveCapacity(values.count)
-
-    for value in values.reversed() {
-        guard seenKeys.insert(key(value)).inserted else { continue }
-        result.append(value)
-    }
-
-    return Array(result.reversed())
-}
-
-struct TimelinePagingState: Equatable {
-    var hasMoreBefore: Bool
-    var hasMoreAfter: Bool
-    var isLoadingBefore: Bool
-    var isLoadingAfter: Bool
-
-    static let empty = TimelinePagingState(
-        hasMoreBefore: false,
-        hasMoreAfter: false,
-        isLoadingBefore: false,
-        isLoadingAfter: false
-    )
-}
-
-private nonisolated final class OffMainCancellationFlag: @unchecked Sendable {
-    private let lock = NSLock()
-    private var cancelled = false
-
-    func cancel() {
-        lock.lock()
-        cancelled = true
-        lock.unlock()
-    }
-
-    func check() throws {
-        lock.lock()
-        let isCancelled = cancelled
-        lock.unlock()
-        if isCancelled {
-            throw CancellationError()
-        }
-    }
-}
-
-/// Tracks ownership of incremental, per-row chat-list enrichment tasks (issue #40).
-///
-/// Single-row chat-list updates spawn one enrichment `Task` per group. Exactly one such task
-/// should "own" a group's slot at a time: a newer update must supersede (coalesce) an in-flight
-/// one, listener teardown / account switch must cancel them all, and a finishing task must
-/// release its slot only if it is still the current owner.
-///
-/// Ownership is keyed by a process-monotonic token that is **never reused** — not even after
-/// `cancelAll()` clears the maps on reload / account switch. That is the crux of the fix: a
-/// per-group counter that reset to its first value on clear would let a stale, already-canceled
-/// task match a *future* task's reused token and erroneously drop the future task's slot,
-/// reintroducing the untracked / uncancellable enrichment work this is meant to prevent.
-struct ChatListRowEnrichmentTracker {
-    private var tasks: [String: Task<Void, Never>] = [:]
-    private var tokens: [String: Int] = [:]
-    private var nextToken: Int = 0
-
-    /// Number of currently tracked (live) tasks. Diagnostic / test helper.
-    var trackedTaskCount: Int { tasks.count }
-
-    /// The current ownership token for `group`, if any. Diagnostic / test helper.
-    func currentToken(forGroup group: String) -> Int? { tokens[group] }
-
-    /// Allocates a globally unique, never-reused ownership token for `group` and cancels any
-    /// task currently owning it. Call before spawning the replacement task.
-    mutating func beginTask(forGroup group: String) -> Int {
-        tasks[group]?.cancel()
-        nextToken += 1
-        let token = nextToken
-        tokens[group] = token
-        return token
-    }
-
-    /// Records `task` as the owner of `group` for `token`. If `token` is no longer current
-    /// (a newer `beginTask` has since run for this group) the late registration is ignored and
-    /// the task canceled, so it cannot clobber a newer owner.
-    mutating func register(task: Task<Void, Never>, forGroup group: String, token: Int) {
-        guard tokens[group] == token else {
-            task.cancel()
-            return
-        }
-        tasks[group] = task
-    }
-
-    /// Releases `group`'s slot iff `token` is still the current owner. A stale token (from an
-    /// older, already-superseded or canceled task) is a no-op, so it can never drop a newer task.
-    mutating func finishTask(forGroup group: String, token: Int) {
-        guard tokens[group] == token else { return }
-        tasks[group] = nil
-        tokens[group] = nil
-    }
-
-    /// Cancels every tracked task and clears all ownership state. The token sequence is
-    /// deliberately **not** reset, so tokens issued after this call stay unique with respect to
-    /// any still-unwinding canceled task.
-    mutating func cancelAll() {
-        for task in tasks.values {
-            task.cancel()
-        }
-        tasks.removeAll()
-        tokens.removeAll()
-    }
-}
-
-nonisolated struct ChatListOrdering {
-    struct UpsertResult {
-        let chats: [ChatItem]
-        let reindexStart: Int?
-    }
-
-    static func sorted(_ chatItems: [ChatItem], pinnedChatIds: Set<String> = []) -> [ChatItem] {
-        chatItems.sorted { lhs, rhs in
-            areInDisplayOrder(lhs, rhs, pinnedChatIds: pinnedChatIds)
-        }
-    }
-
-    static func upserting(
-        _ chat: ChatItem,
-        into chats: [ChatItem],
-        pinnedChatIds: Set<String> = []
-    ) -> [ChatItem] {
-        upsertResult(
-            chat,
-            into: chats,
-            existingIndex: chats.firstIndex(where: { $0.id == chat.id }),
-            pinnedChatIds: pinnedChatIds
-        ).chats
-    }
-
-    static func upserting(
-        _ chat: ChatItem,
-        into chats: [ChatItem],
-        existingIndex: Int?,
-        pinnedChatIds: Set<String> = []
-    ) -> [ChatItem] {
-        upsertResult(chat, into: chats, existingIndex: existingIndex, pinnedChatIds: pinnedChatIds).chats
-    }
-
-    static func upsertResult(
-        _ chat: ChatItem,
-        into chats: [ChatItem],
-        existingIndex: Int?,
-        pinnedChatIds: Set<String> = []
-    ) -> UpsertResult {
-        var result = chats
-        if let index = existingIndex {
-            if canReplaceInPlace(chat, at: index, in: result, pinnedChatIds: pinnedChatIds) {
-                result[index] = chat
-                return UpsertResult(chats: result, reindexStart: nil)
-            }
-            result.remove(at: index)
-            let nextIndex = insertionIndex(for: chat, in: result, pinnedChatIds: pinnedChatIds)
-            result.insert(chat, at: nextIndex)
-            return UpsertResult(chats: result, reindexStart: min(index, nextIndex))
-        }
-
-        let nextIndex = insertionIndex(for: chat, in: result, pinnedChatIds: pinnedChatIds)
-        result.insert(chat, at: nextIndex)
-        return UpsertResult(chats: result, reindexStart: nextIndex)
-    }
-
-    static func mostRecent(in chatItems: [ChatItem]) -> ChatItem? {
-        chatItems.min { lhs, rhs in
-            areInDisplayOrder(lhs, rhs)
-        }
-    }
-
-    static func preservingResolvedMetadata(in chat: ChatItem, from current: ChatItem) -> ChatItem {
-        ChatItem(
-            id: chat.id,
-            title: current.title,
-            subtitle: current.subtitle,
-            preview: chat.preview,
-            previewAttachmentKind: chat.previewAttachmentKind,
-            previewAttribution: chat.previewAttribution,
-            updatedAt: chat.updatedAt,
-            avatarSeed: current.avatarSeed,
-            pictureURL: current.pictureURL,
-            groupImagePayload: current.groupImagePayload,
-            groupImageHashHex: current.groupImageHashHex,
-            unreadCount: chat.unreadCount,
-            hasUnread: chat.hasUnread,
-            manuallyMarkedUnread: chat.manuallyMarkedUnread,
-            unreadMentionCount: chat.unreadMentionCount,
-            isDirect: chat.hasAuthoritativeConversationKind ? chat.isDirect : current.isDirect,
-            hasAuthoritativeConversationKind: chat.hasAuthoritativeConversationKind,
-            muted: chat.muted,
-            mutedUntilMs: chat.mutedUntilMs,
-            leaveRequestPending: chat.leaveRequestPending,
-            latestMessageDelivery: chat.latestMessageDelivery,
-            pendingConfirmation: chat.pendingConfirmation,
-            selfMembership: chat.selfMembership
-        )
-    }
-
-    static func isOlder(_ candidate: ChatItem, than current: ChatItem) -> Bool {
-        guard let currentUpdatedAt = current.updatedAt else { return false }
-        guard let candidateUpdatedAt = candidate.updatedAt else { return true }
-        return candidateUpdatedAt < currentUpdatedAt
-    }
-
-    private static func canReplaceInPlace(
-        _ chat: ChatItem,
-        at index: Int,
-        in chats: [ChatItem],
-        pinnedChatIds: Set<String>
-    ) -> Bool {
-        let previousStillBefore =
-            index == chats.startIndex
-            || !areInDisplayOrder(chat, chats[index - 1], pinnedChatIds: pinnedChatIds)
-        let nextStillAfter: Bool
-        if index == chats.index(before: chats.endIndex) {
-            nextStillAfter = true
-        } else {
-            nextStillAfter = !areInDisplayOrder(chats[index + 1], chat, pinnedChatIds: pinnedChatIds)
-        }
-        return previousStillBefore && nextStillAfter
-    }
-
-    private static func insertionIndex(
-        for chat: ChatItem,
-        in chats: [ChatItem],
-        pinnedChatIds: Set<String>
-    ) -> Int {
-        var lowerBound = chats.startIndex
-        var upperBound = chats.endIndex
-        while lowerBound < upperBound {
-            let middle = lowerBound + (upperBound - lowerBound) / 2
-            if areInDisplayOrder(chats[middle], chat, pinnedChatIds: pinnedChatIds) {
-                lowerBound = middle + 1
-            } else {
-                upperBound = middle
-            }
-        }
-        return lowerBound
-    }
-
-    private static func areInDisplayOrder(
-        _ lhs: ChatItem,
-        _ rhs: ChatItem,
-        pinnedChatIds: Set<String> = []
-    ) -> Bool {
-        let lhsPinned = pinnedChatIds.contains(lhs.id)
-        let rhsPinned = pinnedChatIds.contains(rhs.id)
-        if lhsPinned != rhsPinned { return lhsPinned }
-
-        switch (lhs.updatedAt, rhs.updatedAt) {
-        case (let left?, let right?) where left != right:
-            return left > right
-        case (_?, nil):
-            return true
-        case (nil, _?):
-            return false
-        default:
-            return lhs.title.localizedCaseInsensitiveCompare(rhs.title) == .orderedAscending
-        }
-    }
-}
-
-@MainActor
-@Observable
-final class MediaDownloadStateStore {
-    private(set) var state: MediaDownloadState = .idle
-
-    var shouldStartAutomaticDownload: Bool {
-        if case .idle = state {
-            return true
-        }
-        return false
-    }
-
-    func update(_ newState: MediaDownloadState) {
-        guard state != newState else { return }
-        state = newState
-    }
-}
-
-@MainActor
-@Observable
-final class MessageTimelineStore {
-    struct ProjectionApplyResult: Equatable {
-        let didChange: Bool
-        let didRemoveMessages: Bool
-        let didTrimOlderMessages: Bool
-        let didChangeMediaAttachments: Bool
-    }
-
-    private(set) var messages: [MessageItem]
-    private(set) var messageIDs: [String]
-    /// Day-grouped rendering projection. Rebuilt only when the materialized message window or
-    /// its day/locale context changes, never from an unrelated `ConversationView.body` pass.
-    private(set) var displayItems: [TimelineMessageDisplayItem]
-    @ObservationIgnored private var displayReferenceDate: Date
-    @ObservationIgnored private var displayCalendar: Calendar
-    @ObservationIgnored private var displayLocale: Locale
-    /// O(1) id → message map for non-UI lookups (`timelineMessage(groupIdHex:messageId:)`).
-    /// `@ObservationIgnored`: it is never read from a view body — only `messages`/`messageIDs`
-    /// drive rendering — so it must not enlarge a view's observation set. The store owns this
-    /// (and `messageIDs`) so callers don't maintain parallel per-chat dictionaries.
-    @ObservationIgnored private(set) var lookup: [String: MessageItem]
-    @ObservationIgnored private var indexById: [String: Int]
-    /// Unedited chat targets keyed by message id. Rendered rows in `messages`/`lookup` are derived
-    /// from these bases plus the newest valid kind-1009 candidate per materialized target.
-    @ObservationIgnored private var baseMessagesById: [String: MessageItem] = [:]
-    /// Active kind-1009 edit candidates keyed by edit-event id. Retained across trim and
-    /// authoritative window replaces that omit edit records; bounded by `windowLimit`.
-    @ObservationIgnored private var editCandidatesById: [String: MessageEditOverlay] = [:]
-    /// The same candidates grouped by target id, rebuilt whenever the canonical id map changes.
-    /// Rendering a full window therefore visits only each row's candidates instead of rescanning
-    /// every retained edit for every row.
-    @ObservationIgnored private var editCandidatesByTargetId: [String: [MessageEditOverlay]] = [:]
-    /// Candidates inspected by the most recent full render pass. Diagnostic / test helper.
-    @ObservationIgnored private(set) var lastRenderEditCandidateVisitCount = 0
-    /// Earliest sort key of a media-changing upsert suppressed from a detached window.
-    /// Consumed when an authoritative replace first retains media at/after this key.
-    @ObservationIgnored private var pendingSuppressedMediaSortKey: TimelineSortKey?
-    private(set) var isLoaded: Bool
-
-    init(
-        messages: [MessageItem] = [],
-        isLoaded: Bool = false,
-        displayReferenceDate: Date = Date(),
-        displayCalendar: Calendar = .autoupdatingCurrent,
-        displayLocale: Locale = AppLanguage.currentLocale
-    ) {
-        let messages = Self.deduplicatedMessages(messages)
-        self.messages = messages
-        self.messageIDs = messages.map(\.id)
-        self.displayItems = TimelineMessageDisplayItem.make(
-            from: messages,
-            now: displayReferenceDate,
-            calendar: displayCalendar,
-            locale: displayLocale
-        )
-        self.displayReferenceDate = displayReferenceDate
-        self.displayCalendar = displayCalendar
-        self.displayLocale = displayLocale
-        self.lookup = Dictionary(messages.map { ($0.id, $0) }, uniquingKeysWith: { _, new in new })
-        self.indexById = Dictionary(
-            messages.enumerated().map { ($0.element.id, $0.offset) },
-            uniquingKeysWith: { _, new in new }
-        )
-        self.baseMessagesById = Dictionary(messages.map { ($0.id, $0) }, uniquingKeysWith: { _, new in new })
-        self.isLoaded = isLoaded
-        #if DEBUG
-            displayItemsBuildCount = 1
-        #endif
-    }
-
-    private static func deduplicatedMessages(_ messages: [MessageItem]) -> [MessageItem] {
-        uniquedLastWins(messages, by: \.id)
-    }
-
-    static func loaded(with messages: [MessageItem]) -> MessageTimelineStore {
-        MessageTimelineStore(messages: messages, isLoaded: true)
-    }
-
-    @discardableResult
-    func replace(
-        with messages: [MessageItem],
-        editMutations: [MessageEditMutation] = [],
-        windowLimit: Int? = nil
-    ) -> Bool {
-        let priorLookup = lookup
-        let bases = Self.deduplicatedMessages(messages)
-        self.messages = bases
-        baseMessagesById = Dictionary(bases.map { ($0.id, $0) }, uniquingKeysWith: { _, new in new })
-        rebuildIndexes()
-        applyEditMutations(editMutations)
-        pruneInvalidSenderCandidatesForMaterializedTargets()
-        pruneEditCandidates(windowLimit: windowLimit ?? max(bases.count, 1))
-        _ = recomputeAllRenderedMessages()
-        rebuildDisplayItems()
-        self.isLoaded = true
-        return consumeMediaAttachmentInvalidation(afterAuthoritativeReplace: bases, priorLookup: priorLookup)
-    }
-
-    func clear() {
-        messages = []
-        messageIDs = []
-        displayItems = []
-        lookup = [:]
-        indexById = [:]
-        baseMessagesById = [:]
-        replaceEditCandidates(with: [:])
-        pendingSuppressedMediaSortKey = nil
-        isLoaded = false
-    }
-
-    func containsMessage(id: String) -> Bool {
-        lookup[id] != nil
-    }
-
-    func applyProjection(
-        upserts: [MessageItem],
-        removals removalIds: Set<String>,
-        editMutations: [MessageEditMutation] = [],
-        anchoredToNewest: Bool,
-        windowLimit: Int
-    ) -> ProjectionApplyResult {
-        var didChange = false
-        var didRemoveMessages = false
-
-        // Compare media-relevant upserts with the final lookup after trimming so suppressed or
-        // immediately-evicted rows do not trigger cache invalidation.
-        var mediaAttachmentsBeforeUpserts: [String: [MessageMediaAttachment]] = [:]
-        for item in upserts {
-            guard mediaAttachmentsBeforeUpserts[item.id] == nil else { continue }
-            let previous = lookup[item.id]?.mediaAttachments ?? []
-            guard !previous.isEmpty || !item.mediaAttachments.isEmpty else { continue }
-            mediaAttachmentsBeforeUpserts[item.id] = previous
-        }
-
-        if !removalIds.isEmpty {
-            let originalCount = messages.count
-            let removedMessageIds = Set(messages.filter { removalIds.contains($0.id) }.map(\.id))
-            messages.removeAll { removalIds.contains($0.id) }
-            didRemoveMessages = messages.count != originalCount
-            for removedId in removedMessageIds {
-                baseMessagesById.removeValue(forKey: removedId)
-            }
-            purgeEditCandidates(forRemovalIds: removalIds)
-            if didRemoveMessages {
-                didChange = true
-                rebuildIndexes()
-            }
-        }
-
-        // Recompute the head *after* removals: a delta that drops the current newest row lowers
-        // the detached window's real head, and an upsert newer than that post-removal head must
-        // not grow a new head into a scrolled-back window.
-        let newestKey = messages.last.map(TimelineSortKey.init)
-
-        for item in upserts {
-            if let existingIndex = indexById[item.id] {
-                baseMessagesById[item.id] = item
-                guard messages[existingIndex] != item else { continue }
-                if TimelineSortKey(messages[existingIndex]) == TimelineSortKey(item) {
-                    messages[existingIndex] = item
-                    lookup[item.id] = item
-                    didChange = true
-                    continue
-                }
-                removeMessage(at: existingIndex)
-                insertMessage(item, at: insertionIndex(for: item))
-                didChange = true
-                continue
-            }
-
-            let isInsideDetachedWindow = newestKey.map { TimelineSortKey(item) <= $0 } ?? false
-            guard anchoredToNewest || isInsideDetachedWindow else { continue }
-            baseMessagesById[item.id] = item
-            insertMessage(item, at: insertionIndex(for: item))
-            didChange = true
-        }
-
-        applyEditMutations(editMutations)
-
-        var didTrimOlderMessages = false
-        if messages.count > windowLimit {
-            trimOldestMessages(count: messages.count - windowLimit)
-            didTrimOlderMessages = true
-            didChange = true
-        }
-
-        pruneInvalidSenderCandidatesForMaterializedTargets()
-        pruneEditCandidates(windowLimit: windowLimit)
-        if recomputeAllRenderedMessages() {
-            didChange = true
-        }
-
-        if didChange {
-            rebuildDisplayItems()
-            isLoaded = true
-        }
-        let didChangeMediaAttachments = mediaAttachmentsBeforeUpserts.contains { id, previous in
-            guard let retained = lookup[id] else { return false }
-            return previous != retained.mediaAttachments
-        }
-        if didChangeMediaAttachments {
-            pendingSuppressedMediaSortKey = nil
-        } else {
-            for item in upserts {
-                guard let previous = mediaAttachmentsBeforeUpserts[item.id] else { continue }
-                guard lookup[item.id] == nil else { continue }
-                guard !previous.isEmpty || !item.mediaAttachments.isEmpty else { continue }
-                guard previous != item.mediaAttachments else { continue }
-                recordPendingSuppressedMediaSortKey(for: item)
-            }
-        }
-        return ProjectionApplyResult(
-            didChange: didChange,
-            didRemoveMessages: didRemoveMessages,
-            didTrimOlderMessages: didTrimOlderMessages,
-            didChangeMediaAttachments: didChangeMediaAttachments
-        )
-    }
-
-    @discardableResult
-    func relabelSender(accountIdHex: String, nickname: String?) -> Bool {
-        var didChange = false
-        for index in messages.indices {
-            guard messages[index].senderAccountIdHex == accountIdHex,
-                let relabeled = messages[index].applyingSenderNickname(nickname)
-            else { continue }
-            messages[index] = relabeled
-            lookup[relabeled.id] = relabeled
-            didChange = true
-        }
-        for (id, base) in baseMessagesById where base.senderAccountIdHex == accountIdHex {
-            guard let relabeled = base.applyingSenderNickname(nickname) else { continue }
-            baseMessagesById[id] = relabeled
-            didChange = true
-        }
-        if relabelReplyQuotes(ofSender: accountIdHex) {
-            didChange = true
-        }
-        guard didChange else { return false }
-        rebuildDisplayItems()
-        return true
-    }
-
-    /// Rewrites every mention of one person in this materialized window, the counterpart of
-    /// `relabelSender` for the labels baked into a bubble's body. Rows that do not mention them
-    /// are skipped without allocating, so the cost is proportional to the mentions actually on
-    /// screen rather than to the window size.
-    ///
-    /// Both the rendered rows and the pre-edit bases are relabeled: an edited row renders from its
-    /// base, and a base left with the stale label would resurface it on the next recompute.
-    ///
-    /// An edited row cannot be relabeled directly — it carries no Markdown (`applyingEdit` drops
-    /// it) and resolves its mentions into plain text from the base's `mentionNames` — so the rows
-    /// whose base changed are re-rendered through the ordinary edit-overlay path afterwards.
-    /// Without that, an edited bubble kept the old label until the next full recomputation.
-    @discardableResult
-    func relabelMention(bech32: String, name: String?) -> Bool {
-        var didChange = false
-        for index in messages.indices {
-            guard let relabeled = messages[index].applyingMentionLabel(bech32: bech32, name: name) else { continue }
-            messages[index] = relabeled
-            lookup[relabeled.id] = relabeled
-            didChange = true
-        }
-        var relabeledBaseIds: Set<String> = []
-        for (id, base) in baseMessagesById {
-            guard let relabeled = base.applyingMentionLabel(bech32: bech32, name: name) else { continue }
-            baseMessagesById[id] = relabeled
-            relabeledBaseIds.insert(id)
-            didChange = true
-        }
-        if recomputeRenderedMessages(withIds: relabeledBaseIds) {
-            didChange = true
-        }
-        guard didChange else { return false }
-        rebuildDisplayItems()
-        return true
-    }
-
-    private func relabelReplyQuotes(ofSender accountIdHex: String) -> Bool {
-        var didChange = false
-        for index in messages.indices {
-            guard let targetId = messages[index].replyContext?.targetMessageId,
-                let target = lookup[targetId],
-                target.senderAccountIdHex == accountIdHex,
-                let relabeled = messages[index].applyingReplyContextSenderName(target.senderName)
-            else { continue }
-            messages[index] = relabeled
-            lookup[relabeled.id] = relabeled
-            didChange = true
-        }
-        for (id, base) in baseMessagesById {
-            guard let targetId = base.replyContext?.targetMessageId,
-                let target = lookup[targetId],
-                target.senderAccountIdHex == accountIdHex,
-                let relabeled = base.applyingReplyContextSenderName(target.senderName)
-            else { continue }
-            baseMessagesById[id] = relabeled
-            didChange = true
-        }
-        return didChange
-    }
-
-    /// Refreshes date-sensitive day labels after a calendar rollover or locale change. Calls made
-    /// repeatedly within the same calendar day and locale are intentionally O(1).
-    func refreshDisplayItems(
-        referenceDate: Date,
-        calendar: Calendar = .autoupdatingCurrent,
-        locale: Locale = AppLanguage.currentLocale
-    ) {
-        let dayChanged = !calendar.isDate(displayReferenceDate, inSameDayAs: referenceDate)
-        guard dayChanged || displayCalendar != calendar || displayLocale != locale else { return }
-
-        displayReferenceDate = referenceDate
-        displayCalendar = calendar
-        displayLocale = locale
-        rebuildDisplayItems()
-    }
-
-    private func rebuildDisplayItems() {
-        displayItems = TimelineMessageDisplayItem.make(
-            from: messages,
-            now: displayReferenceDate,
-            calendar: displayCalendar,
-            locale: displayLocale
-        )
-        #if DEBUG
-            displayItemsBuildCount += 1
-        #endif
-    }
-
-    #if DEBUG
-        /// Test-only instrumentation proving unrelated view reads do not re-derive day grouping.
-        @ObservationIgnored private(set) var displayItemsBuildCount = 0
-    #endif
-
-    private func applyEditMutations(_ mutations: [MessageEditMutation]) {
-        guard !mutations.isEmpty else { return }
-        for mutation in mutations {
-            switch mutation {
-            case .upsert(let overlay):
-                editCandidatesById[overlay.editMessageIdHex] = overlay
-            case .retract(let editMessageIdHex):
-                editCandidatesById.removeValue(forKey: editMessageIdHex)
-            }
-        }
-        rebuildEditCandidateTargetIndex()
-    }
-
-    @discardableResult
-    private func recomputeAllRenderedMessages() -> Bool {
-        var didChange = false
-        var candidateVisitCount = 0
-        for index in messages.indices {
-            let messageId = messages[index].id
-            guard let base = baseMessagesById[messageId] else { continue }
-            let renderResult = renderedMessage(from: base)
-            candidateVisitCount += renderResult.candidateVisitCount
-            let rendered = renderResult.message
-            guard messages[index] != rendered else { continue }
-            messages[index] = rendered
-            lookup[messageId] = rendered
-            didChange = true
-        }
-        lastRenderEditCandidateVisitCount = candidateVisitCount
-        return didChange
-    }
-
-    /// Re-renders only the rows whose base was just mutated, through the same edit-overlay path
-    /// `recomputeAllRenderedMessages` uses, so a targeted base mutation can never leave a rendered
-    /// row derived from the old base. Visiting only the changed ids keeps the cost proportional to
-    /// the mutation instead of to the window, and rows that render identically are left untouched.
-    ///
-    /// `lastRenderEditCandidateVisitCount` is deliberately not written here: it reports the most
-    /// recent *full* render pass, and a partial count would misreport it.
-    private func recomputeRenderedMessages(withIds ids: Set<String>) -> Bool {
-        guard !ids.isEmpty else { return false }
-        var didChange = false
-        for id in ids {
-            guard let index = indexById[id], let base = baseMessagesById[id] else { continue }
-            let rendered = renderedMessage(from: base).message
-            guard messages[index] != rendered else { continue }
-            messages[index] = rendered
-            lookup[id] = rendered
-            didChange = true
-        }
-        return didChange
-    }
-
-    private func renderedMessage(from base: MessageItem) -> (message: MessageItem, candidateVisitCount: Int) {
-        let result = effectiveEdit(for: base.id, base: base)
-        guard let edit = result.edit else { return (base, result.candidateVisitCount) }
-        return (base.applyingEdit(plaintext: edit.plaintext), result.candidateVisitCount)
-    }
-
-    private func effectiveEdit(
-        for targetId: String,
-        base: MessageItem
-    ) -> (edit: MessageEditOverlay?, candidateVisitCount: Int) {
-        guard isValidEditTarget(base, editSender: base.senderAccountIdHex),
-            let candidates = editCandidatesByTargetId[targetId]
-        else { return (nil, 0) }
-        var winner: MessageEditOverlay?
-        var candidateVisitCount = 0
-        for candidate in candidates {
-            candidateVisitCount += 1
-            guard candidate.sender == base.senderAccountIdHex else { continue }
-            if MessageEditOverlay.shouldPrefer(candidate, over: winner) {
-                winner = candidate
-            }
-        }
-        return (winner, candidateVisitCount)
-    }
-
-    /// The ordered edit history for `targetId` — oldest first: the original, then each accepted
-    /// edit by the target's own author. Empty when the message was never edited.
-    func editHistory(forTarget targetId: String) -> [MessageEditVersion] {
-        guard let base = baseMessagesById[targetId],
-            isValidEditTarget(base, editSender: base.senderAccountIdHex)
-        else { return [] }
-        let edits = (editCandidatesByTargetId[targetId] ?? [])
-            .filter { $0.sender == base.senderAccountIdHex }
-            .sorted { MessageEditOverlay.shouldPrefer($1, over: $0) }
-        guard !edits.isEmpty else { return [] }
-        var versions = [
-            MessageEditVersion(
-                id: base.id,
-                text: MentionDisplayResolver.resolve(in: base.wireBody, mentionNames: base.mentionNames),
-                date: Date(timeIntervalSince1970: TimeInterval(base.timelineAt)),
-                isOriginal: true
-            )
-        ]
-        versions += edits.map { edit in
-            MessageEditVersion(
-                id: edit.editMessageIdHex,
-                text: MentionDisplayResolver.resolve(in: edit.plaintext, mentionNames: base.mentionNames),
-                date: Date(timeIntervalSince1970: TimeInterval(edit.timelineAt)),
-                isOriginal: false
-            )
-        }
-        return versions
-    }
-
-    private func purgeEditCandidates(forRemovalIds removalIds: Set<String>) {
-        guard !removalIds.isEmpty, !editCandidatesById.isEmpty else { return }
-        replaceEditCandidates(
-            with: editCandidatesById.filter { _, candidate in
-                !removalIds.contains(candidate.editMessageIdHex)
-                    && !removalIds.contains(candidate.targetMessageIdHex)
-            }
-        )
-    }
-
-    private func pruneInvalidSenderCandidatesForMaterializedTargets() {
-        guard !editCandidatesById.isEmpty else { return }
-        replaceEditCandidates(
-            with: editCandidatesById.filter { _, candidate in
-                guard let base = baseMessagesById[candidate.targetMessageIdHex] else { return true }
-                return candidate.sender == base.senderAccountIdHex
-            }
-        )
-    }
-
-    private func pruneEditCandidates(windowLimit limit: Int) {
-        guard limit > 0, editCandidatesById.count > limit else { return }
-
-        typealias CandidateEntry = (key: String, value: MessageEditOverlay)
-
-        func sortNewestFirst(_ lhs: CandidateEntry, _ rhs: CandidateEntry) -> Bool {
-            if lhs.value.timelineAt != rhs.value.timelineAt {
-                return lhs.value.timelineAt > rhs.value.timelineAt
-            }
-            return lhs.value.editMessageIdHex > rhs.value.editMessageIdHex
-        }
-
-        // Preserve the current winner for each visible target. For unresolved targets,
-        // also preserve one representative per target/sender pair before using spare
-        // capacity for history. Repeated forged edits from one peer therefore cannot
-        // evict a pending candidate from the target's actual author.
-        var materializedWinnerIds = Set<String>()
-        for (targetId, base) in baseMessagesById {
-            if let winner = effectiveEdit(for: targetId, base: base).edit {
-                materializedWinnerIds.insert(winner.editMessageIdHex)
-            }
-        }
-
-        var pendingRepresentativesByTarget = [String: [String: MessageEditOverlay]]()
-        for candidate in editCandidatesById.values where baseMessagesById[candidate.targetMessageIdHex] == nil {
-            let existing = pendingRepresentativesByTarget[candidate.targetMessageIdHex]?[candidate.sender]
-            if MessageEditOverlay.shouldPrefer(candidate, over: existing) {
-                pendingRepresentativesByTarget[candidate.targetMessageIdHex, default: [:]][candidate.sender] = candidate
-            }
-        }
-        let pendingRepresentativeIds = Set(
-            pendingRepresentativesByTarget.values
-                .flatMap(\.values)
-                .map(\.editMessageIdHex)
-        )
-
-        let materializedWinners =
-            editCandidatesById
-            .filter { materializedWinnerIds.contains($0.key) }
-            .sorted(by: sortNewestFirst)
-            .prefix(limit)
-        var kept = Dictionary(
-            uniqueKeysWithValues: materializedWinners.map { ($0.key, $0.value) }
-        )
-
-        let pendingSlots = limit - kept.count
-        if pendingSlots > 0 {
-            let pendingRepresentatives =
-                editCandidatesById
-                .filter { pendingRepresentativeIds.contains($0.key) }
-                .sorted(by: sortNewestFirst)
-                .prefix(pendingSlots)
-            for candidate in pendingRepresentatives {
-                kept[candidate.key] = candidate.value
-            }
-        }
-
-        let remainingSlots = limit - kept.count
-        if remainingSlots > 0 {
-            let remaining =
-                editCandidatesById
-                .filter { kept[$0.key] == nil }
-                .sorted(by: sortNewestFirst)
-                .prefix(remainingSlots)
-            for candidate in remaining {
-                kept[candidate.key] = candidate.value
-            }
-        }
-        replaceEditCandidates(with: kept)
-    }
-
-    private func replaceEditCandidates(with candidates: [String: MessageEditOverlay]) {
-        editCandidatesById = candidates
-        rebuildEditCandidateTargetIndex()
-    }
-
-    private func rebuildEditCandidateTargetIndex() {
-        editCandidatesByTargetId = Dictionary(
-            grouping: editCandidatesById.values,
-            by: { $0.targetMessageIdHex }
-        )
-    }
-
-    private func isValidEditTarget(_ message: MessageItem, editSender: String) -> Bool {
-        message.timelineKind == 9
-            && message.presentation == .chat
-            && message.senderAccountIdHex == editSender
-            && !message.isDeleted
-            && message.invalidationStatus == nil
-    }
-
-    private func rebuildIndexes() {
-        messages = Self.deduplicatedMessages(messages)
-        messageIDs = messages.map(\.id)
-        lookup = Dictionary(messages.map { ($0.id, $0) }, uniquingKeysWith: { _, new in new })
-        indexById = Dictionary(
-            messages.enumerated().map { ($0.element.id, $0.offset) },
-            uniquingKeysWith: { _, new in new }
-        )
-    }
-
-    private func trimOldestMessages(count trimCount: Int) {
-        guard trimCount > 0 else { return }
-        let trimmedIDs = messages.prefix(trimCount).map(\.id)
-        messages.removeFirst(trimCount)
-        messageIDs.removeFirst(trimCount)
-        for id in trimmedIDs {
-            lookup[id] = nil
-            indexById[id] = nil
-            baseMessagesById.removeValue(forKey: id)
-        }
-        reindexMessages(startingAt: 0)
-    }
-
-    private func insertMessage(_ item: MessageItem, at index: Int) {
-        messages.insert(item, at: index)
-        messageIDs.insert(item.id, at: index)
-        lookup[item.id] = item
-        reindexMessages(startingAt: index)
-    }
-
-    private func removeMessage(at index: Int) {
-        let removed = messages.remove(at: index)
-        messageIDs.remove(at: index)
-        lookup[removed.id] = nil
-        indexById[removed.id] = nil
-        reindexMessages(startingAt: index)
-    }
-
-    private func reindexMessages(startingAt startIndex: Int) {
-        guard startIndex < messages.endIndex else { return }
-        for index in startIndex..<messages.endIndex {
-            indexById[messages[index].id] = index
-        }
-    }
-
-    private func recordPendingSuppressedMediaSortKey(for item: MessageItem) {
-        guard !item.mediaAttachments.isEmpty else { return }
-        let key = TimelineSortKey(item)
-        if let existing = pendingSuppressedMediaSortKey, key >= existing {
-            return
-        }
-        pendingSuppressedMediaSortKey = key
-    }
-
-    private func consumeMediaAttachmentInvalidation(
-        afterAuthoritativeReplace bases: [MessageItem],
-        priorLookup: [String: MessageItem]
-    ) -> Bool {
-        let didChangeMediaAttachments = bases.contains { item in
-            guard let previous = priorLookup[item.id] else { return false }
-            let previousAttachments = previous.mediaAttachments
-            let nextAttachments = item.mediaAttachments
-            guard !previousAttachments.isEmpty || !nextAttachments.isEmpty else { return false }
-            return previousAttachments != nextAttachments
-        }
-        if didChangeMediaAttachments {
-            pendingSuppressedMediaSortKey = nil
-            return true
-        }
-        guard let watermark = pendingSuppressedMediaSortKey else { return false }
-        guard
-            bases.contains(where: { item in
-                !item.mediaAttachments.isEmpty && TimelineSortKey(item) >= watermark
-            })
-        else {
-            return false
-        }
-        pendingSuppressedMediaSortKey = nil
-        return true
-    }
-
-    private func insertionIndex(for item: MessageItem) -> Int {
-        let key = TimelineSortKey(item)
-        var lowerBound = messages.startIndex
-        var upperBound = messages.endIndex
-        while lowerBound < upperBound {
-            let middle = lowerBound + (upperBound - lowerBound) / 2
-            if TimelineSortKey(messages[middle]) < key {
-                lowerBound = middle + 1
-            } else {
-                upperBound = middle
-            }
-        }
-        return lowerBound
-    }
-
-    private struct TimelineSortKey: Comparable {
-        let timelineAt: UInt64
-        let id: String
-
-        init(_ message: MessageItem) {
-            timelineAt = message.timelineAt
-            id = message.id
-        }
-
-        static func < (lhs: TimelineSortKey, rhs: TimelineSortKey) -> Bool {
-            if lhs.timelineAt != rhs.timelineAt {
-                return lhs.timelineAt < rhs.timelineAt
-            }
-            return lhs.id < rhs.id
-        }
-    }
-}
 
 @MainActor
 @Observable
@@ -1709,7 +751,7 @@ final class WorkspaceState {
     /// at a time must cost one re-projection, not one per member.
     static let peerProfileReprojectionDebounceNanoseconds: UInt64 = 250_000_000
     /// Concurrent relay refreshes. Android allows 6; this app also funnels the local
-    /// re-resolution through one serialized FFI queue (`ffiQueue`), so a lower ceiling
+    /// re-resolution through one shared FFI queue (`FFIExecutor.queue`), so a lower ceiling
     /// keeps the timeline hot path clear.
     static let peerProfileRefreshFanout = 4
     /// How many times a still-draining queue may restart the re-projection debounce. Caps the
@@ -1852,16 +894,6 @@ final class WorkspaceState {
         return listenerReconnectDelaysNanoseconds[index]
     }
 
-    /// Dedicated queue for blocking MarmotRuntime FFI calls. The Rust core runs
-    /// synchronously (DB reads, MLS decryption); WorkspaceState is `@MainActor`, so
-    /// calling these directly freezes the UI. We hop them onto this queue and await the
-    /// result on the main actor. UniFFI objects are internally thread-safe.
-    nonisolated static let ffiQueue = DispatchQueue(
-        label: "chat.whitenoise.marmot-ffi",
-        qos: .userInitiated,
-        attributes: .concurrent
-    )
-
     /// Runs the pure timeline record → view-model transformation off the main actor.
     /// This uses the same queue as blocking FFI work because `WorkspaceState` is
     /// `@MainActor` while Markdown/attributed-string/media-JSON mapping can be expensive.
@@ -1872,7 +904,7 @@ final class WorkspaceState {
         mentionNames: MarkdownMentionNames
     ) async -> [MessageItem] {
         await withCheckedContinuation { (continuation: CheckedContinuation<[MessageItem], Never>) in
-            Self.ffiQueue.async {
+            FFIExecutor.queue.async {
                 continuation.resume(
                     returning: MessageItem.timeline(
                         from: page,
@@ -2015,51 +1047,6 @@ final class WorkspaceState {
         var npub: String?
     }
 
-    /// Runs a blocking FFI closure off the main thread and resumes on the caller's actor.
-    nonisolated static func runFFI<T>(
-        _ work: @escaping @Sendable () throws -> T
-    ) async throws -> T {
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<T, Error>) in
-            Self.ffiQueue.async {
-                continuation.resume(with: Result { try work() })
-            }
-        }
-    }
-
-    /// Runs a blocking FFI closure off the main thread and resumes on the caller's actor.
-    nonisolated func runOffMain<T>(
-        _ work: @escaping @Sendable () throws -> T
-    ) async throws -> T {
-        try await Self.runFFI(work)
-    }
-
-    /// Runs blocking FFI off the main thread while exposing cancellation to the GCD closure.
-    /// `Task.checkCancellation()` is only task-local; inside `ffiQueue.async` there is no current
-    /// Swift task, so long synchronous loops must call the supplied checker instead.
-    nonisolated func runOffMainCancellable<T>(
-        _ work: @escaping @Sendable (_ checkCancellation: @escaping @Sendable () throws -> Void) throws -> T
-    ) async throws -> T {
-        let cancellation = OffMainCancellationFlag()
-        return try await withTaskCancellationHandler {
-            try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<T, Error>) in
-                Self.ffiQueue.async {
-                    let checkCancellation: @Sendable () throws -> Void = {
-                        try cancellation.check()
-                    }
-                    continuation.resume(
-                        with: Result {
-                            try checkCancellation()
-                            let value = try work(checkCancellation)
-                            try checkCancellation()
-                            return value
-                        }
-                    )
-                }
-            }
-        } onCancel: {
-            cancellation.cancel()
-        }
-    }
     static var notificationPermissionGuidance: String {
         L10n.string("Open System Settings > Notifications and allow White Noise notifications, then try again.")
     }
@@ -2183,7 +1170,7 @@ final class WorkspaceState {
     }
 
     private static func deduplicatedChats(_ chats: [ChatItem]) -> [ChatItem] {
-        uniquedLastWins(chats, by: \.id)
+        chats.uniquedLastWins(by: \.id)
     }
 
     var activeAccount: AccountItem? {
@@ -2823,7 +1810,7 @@ final class WorkspaceState {
         {
             relayRuntimeConfig = cached.relayTelemetryRuntimeConfig
         } else {
-            let installId = try await runOffMain {
+            let installId = try await FFIExecutor.run {
                 try client.telemetryInstallId()
             }
             guard !Task.isCancelled, observabilityRuntimeGeneration == generation,
@@ -2840,7 +1827,7 @@ final class WorkspaceState {
             else { return }
         }
         if observabilityRuntimeConfiguration?.auditLogTrackerConfig != auditTrackerConfig {
-            _ = try await runOffMain {
+            _ = try await FFIExecutor.run {
                 try client.setAuditLogTrackerConfig(config: auditTrackerConfig)
             }
             guard !Task.isCancelled, observabilityRuntimeGeneration == generation,
@@ -2922,291 +1909,6 @@ extension NotificationSettingsSnapshot {
         self.init(
             localNotificationsEnabled: settings.localNotificationsEnabled
         )
-    }
-}
-
-struct LocalNotificationRequest: Equatable {
-    let identifier: String
-    let title: String
-    let body: String
-    let threadIdentifier: String
-    let userInfo: [String: String]
-}
-
-@MainActor
-protocol LocalNotificationCenter: AnyObject {
-    func authorizationStatus() async -> LocalNotificationAuthorizationStatus
-    func requestAuthorization() async throws -> LocalNotificationAuthorizationStatus
-    func post(_ notification: LocalNotificationRequest) async throws
-    func setResponseHandler(_ handler: @escaping @MainActor ([String: String]) -> Void)
-}
-
-@MainActor
-final class MacLocalNotificationCenter: NSObject, LocalNotificationCenter, UNUserNotificationCenterDelegate {
-    private let center: UNUserNotificationCenter
-    private var responseHandler: (@MainActor ([String: String]) -> Void)?
-
-    init(center: UNUserNotificationCenter = .current()) {
-        self.center = center
-        super.init()
-        center.delegate = self
-    }
-
-    func authorizationStatus() async -> LocalNotificationAuthorizationStatus {
-        await currentSettings().authorizationStatus.localNotificationStatus
-    }
-
-    func requestAuthorization() async throws -> LocalNotificationAuthorizationStatus {
-        _ = try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Bool, Error>) in
-            center.requestAuthorization(options: [.alert, .sound, .badge]) { granted, error in
-                if let error {
-                    continuation.resume(throwing: error)
-                } else {
-                    continuation.resume(returning: granted)
-                }
-            }
-        }
-        return await authorizationStatus()
-    }
-
-    func post(_ notification: LocalNotificationRequest) async throws {
-        let content = UNMutableNotificationContent()
-        content.title = notification.title
-        content.body = notification.body
-        content.sound = .default
-        content.threadIdentifier = notification.threadIdentifier
-        content.userInfo = notification.userInfo
-
-        let request = UNNotificationRequest(
-            identifier: notification.identifier,
-            content: content,
-            trigger: nil
-        )
-
-        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
-            center.add(request) { error in
-                if let error {
-                    continuation.resume(throwing: error)
-                } else {
-                    continuation.resume()
-                }
-            }
-        }
-    }
-
-    func setResponseHandler(_ handler: @escaping @MainActor ([String: String]) -> Void) {
-        responseHandler = handler
-    }
-
-    nonisolated func userNotificationCenter(
-        _ center: UNUserNotificationCenter,
-        didReceive response: UNNotificationResponse,
-        withCompletionHandler completionHandler: @escaping () -> Void
-    ) {
-        let userInfo = response.notification.request.content.userInfo.reduce(into: [String: String]()) {
-            result, element in
-            guard let key = element.key as? String else { return }
-            if let value = element.value as? String {
-                result[key] = value
-            }
-        }
-
-        Task { @MainActor [weak self] in
-            self?.responseHandler?(userInfo)
-            completionHandler()
-        }
-    }
-
-    nonisolated func userNotificationCenter(
-        _ center: UNUserNotificationCenter,
-        willPresent notification: UNNotification,
-        withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void
-    ) {
-        completionHandler([.banner, .list, .sound])
-    }
-
-    private func currentSettings() async -> UNNotificationSettings {
-        await withCheckedContinuation { continuation in
-            center.getNotificationSettings { settings in
-                continuation.resume(returning: settings)
-            }
-        }
-    }
-}
-
-extension UNAuthorizationStatus {
-    var localNotificationStatus: LocalNotificationAuthorizationStatus {
-        switch self {
-        case .notDetermined:
-            .notDetermined
-        case .denied:
-            .denied
-        case .authorized:
-            .authorized
-        case .provisional:
-            .provisional
-        case .ephemeral:
-            .ephemeral
-        @unknown default:
-            .denied
-        }
-    }
-}
-
-struct GroupImageSearchResult: Identifiable, Hashable, Sendable {
-    let id: String
-    let title: String
-    let imageURL: String
-    let thumbnailURL: String?
-    let creator: String?
-    let license: String?
-    let attribution: String?
-    let sourceURL: String?
-    let width: Int?
-    let height: Int?
-
-    var dimension: String? {
-        guard let width, let height, width > 0, height > 0 else { return nil }
-        return "\(width)x\(height)"
-    }
-
-    /// The URL used to render the search-result thumbnail. Only the
-    /// Openverse-proxied `thumbnailURL` is used; the arbitrary origin
-    /// `imageURL` is never fetched for previews (whitenoise-mac#315), so a
-    /// result without a usable thumbnail renders the placeholder instead.
-    var previewURL: URL? {
-        guard let thumbnailURL = thumbnailURL?.nilIfBlank else { return nil }
-        return URL(string: thumbnailURL)
-    }
-
-    var creditLine: String {
-        let creatorText = creator?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let licenseText = license?.trimmingCharacters(in: .whitespacesAndNewlines)
-
-        switch (creatorText?.isEmpty == false ? creatorText : nil, licenseText?.isEmpty == false ? licenseText : nil) {
-        case (let creator?, let license?):
-            return "\(creator) · \(license.uppercased())"
-        case (let creator?, nil):
-            return creator
-        case (nil, let license?):
-            return license.uppercased()
-        default:
-            return L10n.string("Openverse")
-        }
-    }
-}
-
-protocol GroupImageSearchClient {
-    func searchImages(query: String) async throws -> [GroupImageSearchResult]
-}
-
-protocol GroupImageSourceLoading {
-    func data(for url: URL) async -> Data?
-}
-
-struct SecureGroupImageSourceLoader: GroupImageSourceLoading {
-    func data(for url: URL) async -> Data? {
-        await RemoteImageLoader.shared.data(for: url)
-    }
-}
-
-struct OpenverseGroupImageSearchClient: GroupImageSearchClient, Sendable {
-    private let endpoint = URL(string: "https://api.openverse.org/v1/images/")!
-
-    func searchImages(query: String) async throws -> [GroupImageSearchResult] {
-        var components = URLComponents(url: endpoint, resolvingAgainstBaseURL: false)
-        components?.queryItems = [
-            URLQueryItem(name: "q", value: query),
-            // Openverse caps anonymous (unauthenticated) requests at page_size 20 and
-            // rejects anything larger with HTTP 401, so stay at the anonymous ceiling.
-            URLQueryItem(name: "page_size", value: "20"),
-            URLQueryItem(name: "mature", value: "false"),
-        ]
-
-        guard let url = components?.url else { throw GroupImageSearchError.invalidURL }
-
-        var request = URLRequest(url: url)
-        request.setValue("WhiteNoiseMac/1.0", forHTTPHeaderField: "User-Agent")
-
-        let (data, response) = try await URLSession.shared.data(for: request)
-        guard let httpResponse = response as? HTTPURLResponse else {
-            throw GroupImageSearchError.invalidResponse
-        }
-        guard (200..<300).contains(httpResponse.statusCode) else {
-            throw GroupImageSearchError.requestFailed(statusCode: httpResponse.statusCode)
-        }
-
-        let decoded = try JSONDecoder().decode(OpenverseImageSearchResponse.self, from: data)
-        return decoded.results.compactMap(\.groupImageSearchResult)
-    }
-}
-
-struct OpenverseImageSearchResponse: Decodable {
-    let results: [OpenverseImageRecord]
-}
-
-struct OpenverseImageRecord: Decodable {
-    let id: String
-    let title: String?
-    let url: String?
-    let thumbnail: String?
-    let creator: String?
-    let license: String?
-    let attribution: String?
-    let foreignLandingURL: String?
-    let width: Int?
-    let height: Int?
-
-    enum CodingKeys: String, CodingKey {
-        case id
-        case title
-        case url
-        case thumbnail
-        case creator
-        case license
-        case attribution
-        case foreignLandingURL = "foreign_landing_url"
-        case width
-        case height
-    }
-
-    var groupImageSearchResult: GroupImageSearchResult? {
-        guard let url = url?.trimmingCharacters(in: .whitespacesAndNewlines),
-            !url.isEmpty,
-            let parsedURL = URL(string: url),
-            ["http", "https"].contains(parsedURL.scheme?.lowercased() ?? "")
-        else { return nil }
-
-        let title = title?.trimmingCharacters(in: .whitespacesAndNewlines)
-        return GroupImageSearchResult(
-            id: id,
-            title: title?.isEmpty == false ? title! : L10n.string("Untitled image"),
-            imageURL: url,
-            thumbnailURL: thumbnail?.nilIfBlank,
-            creator: creator?.nilIfBlank,
-            license: license?.nilIfBlank,
-            attribution: attribution?.nilIfBlank,
-            sourceURL: foreignLandingURL?.nilIfBlank,
-            width: width,
-            height: height
-        )
-    }
-}
-
-enum GroupImageSearchError: LocalizedError {
-    case invalidURL
-    case invalidResponse
-    case requestFailed(statusCode: Int)
-
-    var errorDescription: String? {
-        switch self {
-        case .invalidURL:
-            return L10n.string("Could not build the image search URL.")
-        case .invalidResponse:
-            return L10n.string("The image search service returned an invalid response.")
-        case .requestFailed(let statusCode):
-            return String(format: L10n.string("Image search failed with HTTP status %d."), statusCode)
-        }
     }
 }
